pron-server 站点操作影片上架下架时开启同步开关

+++ zlt-commons/zlt-common-core/src/main/java/com/central/common/model/KpnSiteSign.java
@@ -18,7 +18,7 @@ import java.math.BigDecimal;
 @ApiModel("签到配置")
 public class KpnSiteSign extends SuperEntity {
 	@ApiModelProperty(value = "站点id")
-	private int siteId;
+	private Long siteId;
 
 	@ApiModelProperty(value = "站点编码")
 	private String siteCode;
